backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI and Voice
openai>=1.3.5
anthropic>=0.7.7

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# Mobile Automation
Appium-Python-Client>=3.1.0
selenium>=4.15.2

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Audio processing
pydub>=0.25.1
speechrecognition>=3.10.0

# HTTP client
httpx>=0.25.1
aiohttp>=3.9.1

# Task queue (optional, for async operations)
celery>=5.3.4
redis>=5.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1


